requirements.txt
mcp>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.0
starlette>=0.40.0
anyio>=4.6.0
tenacity>=8.2.0
